refactor: simplify and DRY code
- extract out WinnerCard
- DRY out awardeeShape
- simplify data to one array in scholarships.js
